Add modules to generate thread graphs in GEXF, PNG and DOT
    - Compressed DOT, GEXF and PNG files of threads graphs are at data/graphs
    - Python Modules and corresponding CSV files for the generation of leaf nodes are removed.
